# internal/api/gen/http/openapi3.yaml
openapi: 3.0.3
info:
  title: Enduro API
  version: "1.0"
servers:
- url: http://localhost:9000
paths:
...
  /collection:
    get:
      tags:
      - collection
      summary: list collection
      description: List all stored collections
      operationId: collection#list
      parameters:
      - name: name
        in: query
        allowEmptyValue: true
        schema:
          type: string
          example: Possimus id et cumque.
        example: Earum amet deleniti temporibus ut.
      - name: original_id
        in: query
        allowEmptyValue: true
        schema:
          type: string
          example: Nesciunt aliquam necessitatibus.
        example: Est soluta facere iste.
      - name: transfer_id
        in: query
        allowEmptyValue: true
        schema:
          type: string
          example: 33BA2D68-933B-28F0-E7F4-6D58CB4790FA
          format: uuid
        example: 03F4110E-C4A6-ECD1-6EE0-B074CA2B3425
      - name: aip_id
        in: query
        allowEmptyValue: true
        schema:
          type: string
          example: D27B2E95-C7AA-5857-EDCE-8F7D36361F87
          format: uuid
        example: 911E76DE-4702-489B-4E65-C8542C813C48
      - name: pipeline_id
        in: query
        allowEmptyValue: true
        schema:
          type: string
          example: B6C43FB7-3F0C-5200-5A8D-F20ADE09BC16
          format: uuid
        example: 829FA833-90F0-ADA3-B91B-13E44809B08B
      - name: earliest_created_time
        in: query
        allowEmptyValue: true
        schema:
          type: string
          example: "1980-12-22T21:38:14Z"
          format: date-time
        example: "1982-02-21T18:20:57Z"
      - name: latest_created_time
        in: query
        allowEmptyValue: true
        schema:
          type: string
          example: "2007-06-15T15:07:34Z"
          format: date-time
        example: "1987-07-26T21:48:39Z"
      - name: status
        in: query
        allowEmptyValue: true
        schema:
          type: string
          example: abandoned
          enum:
          - new
          - in progress
          - done
          - error
          - unknown
          - queued
          - pending
          - abandoned
        example: abandoned
      - name: cursor
        in: query
        description: Pagination cursor
        allowEmptyValue: true
        schema:
          type: string
          description: Pagination cursor
          example: Blanditiis quae voluptates facere.
        example: Ut explicabo voluptatem mollitia vitae amet.
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponseBody'
              example:
                items:
                - aip_id: 2C91774A-6D4A-67B6-DC43-6AED849DACC1
                  completed_at: "1997-03-16T23:52:00Z"
                  created_at: "1992-11-05T20:18:50Z"
                  id: 11298461603816854458
                  name: Accusantium dolor molestias.
                  original_id: Nihil recusandae.
                  pipeline_id: 773F4E7F-A748-DF6D-4355-233071D2534A
                  run_id: 74C53540-E974-ABFF-2565-6BF99F9017B2
                  started_at: "2008-06-17T22:56:20Z"
                  status: abandoned
                  transfer_id: FB1E8AC6-4FA4-C883-ED5A-54960E88F5FE
                  workflow_id: 77EB7E77-465C-FCC6-CEC6-11F6C8938D24
                - aip_id: 2C91774A-6D4A-67B6-DC43-6AED849DACC1
                  completed_at: "1997-03-16T23:52:00Z"
                  created_at: "1992-11-05T20:18:50Z"
                  id: 11298461603816854458
                  name: Accusantium dolor molestias.
                  original_id: Nihil recusandae.
                  pipeline_id: 773F4E7F-A748-DF6D-4355-233071D2534A
                  run_id: 74C53540-E974-ABFF-2565-6BF99F9017B2
                  started_at: "2008-06-17T22:56:20Z"
                  status: abandoned
                  transfer_id: FB1E8AC6-4FA4-C883-ED5A-54960E88F5FE
                  workflow_id: 77EB7E77-465C-FCC6-CEC6-11F6C8938D24
                next_cursor: Repellat iusto est libero ut.
  /collection/{id}:
    delete:
      tags:
      - collection
      summary: delete collection
      description: Delete collection by ID
      operationId: collection#delete
      parameters:
      - name: id
        in: path
        description: Identifier of collection to delete
        required: true
        schema:
          type: integer
          description: Identifier of collection to delete
          example: 6637879626105950462
        example: 14732111595377136819
      responses:
        "204":
          description: ""
          content:
            application/json:
              example: {}
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
              example:
                id: 2527394757024271777
                message: Porro velit deserunt vel reiciendis.
    get:
      tags:
      - collection
      summary: show collection
      description: Show collection by ID
      operationId: collection#show
      parameters:
      - name: id
        in: path
        description: Identifier of collection to show
        required: true
        schema:
          type: integer
          description: Identifier of collection to show
          example: 14892226023289384775
        example: 16670709606729443558
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnduroStoredCollection'
              example:
                aip_id: E31B6C92-4A10-2D9D-2B43-B3362900308F
                completed_at: "1985-08-19T04:00:31Z"
                created_at: "2003-04-10T06:28:56Z"
                id: 12449777957699011964
                name: Sint aut veritatis ut quo.
                original_id: Assumenda soluta nihil enim omnis optio voluptatem.
                pipeline_id: EA2D3703-2B9D-E7C0-35B9-DB753C364369
                run_id: F25C441B-B3AA-5871-1077-53EACA084D5A
                started_at: "1976-01-09T01:29:02Z"
                status: abandoned
                transfer_id: E058AD50-5D7F-E9A8-E010-D33044724992
                workflow_id: CEF3CAF7-AB82-083A-2F08-9497BE6B7488
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
              example:
                id: 2168493507218143098
                message: Iure perspiciatis provident placeat id quia.
  /collection/{id}/cancel:
    post:
      tags:
      - collection
      summary: cancel collection
      description: Cancel collection processing by ID
      operationId: collection#cancel
      parameters:
      - name: id
        in: path
        description: Identifier of collection to remove
        required: true
        schema:
          type: integer
          description: Identifier of collection to remove
          example: 17276621612670827821
        example: 10447695724278724003
      responses:
        "200":
          description: ""
          content:
            application/json:
              example: {}
        "400":
          description: ""
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                fault: true
                id: 123abc
                message: parameter 'p' must be an integer
                name: bad_request
                temporary: false
                timeout: false
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
              example:
                id: 6249856683643440275
                message: Recusandae laudantium quidem consequatur ducimus excepturi perferendis.
  /collection/{id}/decision:
    post:
      tags:
      - collection
      summary: decide collection
      description: Make decision for a pending collection by ID
      operationId: collection#decide
      parameters:
      - name: id
        in: path
        description: Identifier of collection to look up
        required: true
        schema:
          type: integer
          description: Identifier of collection to look up
          example: 1435772494281193013
        example: 12475084977748662530
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                option:
                  type: string
                  description: Decision option to proceed with
                  example: Expedita voluptatum.
              example:
                option: Nemo necessitatibus impedit iure id recusandae.
            example:
              option: Explicabo placeat voluptatem aut.
      responses:
        "200":
          description: ""
          content:
            application/json:
              example: {}
        "400":
          description: ""
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                fault: true
                id: 123abc
                message: parameter 'p' must be an integer
                name: bad_request
                temporary: false
                timeout: false
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
              example:
                id: 7529745733278009952
                message: Ut qui voluptatum labore et voluptatem.
...
  /collection/{id}/retry:
    post:
      tags:
      - collection
      summary: retry collection
      description: Retry collection processing by ID
      operationId: collection#retry
      parameters:
      - name: id
        in: path
        description: Identifier of collection to retry
        required: true
        schema:
          type: integer
          description: Identifier of collection to retry
          example: 17938132740731816623
        example: 15523070697073655530
      responses:
        "200":
          description: ""
          content:
            application/json:
              example: {}
        "400":
          description: ""
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                fault: false
                id: 123abc
                message: parameter 'p' must be an integer
                name: bad_request
                temporary: true
                timeout: false
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
              example:
                id: 10364407959241353442
                message: Nihil eius fugit natus iste.
...
components:
  schemas:
...
    ListResponseBody:
      type: object
      properties:
        items:
          $ref: '#/components/schemas/EnduroStoredCollectionCollection'
        next_cursor:
          type: string
          example: Consectetur sint aperiam suscipit nisi quos.
      example:
        items:
        - aip_id: 2C91774A-6D4A-67B6-DC43-6AED849DACC1
          completed_at: "1997-03-16T23:52:00Z"
          created_at: "1992-11-05T20:18:50Z"
          id: 11298461603816854458
          name: Accusantium dolor molestias.
          original_id: Nihil recusandae.
          pipeline_id: 773F4E7F-A748-DF6D-4355-233071D2534A
          run_id: 74C53540-E974-ABFF-2565-6BF99F9017B2
          started_at: "2008-06-17T22:56:20Z"
          status: abandoned
          transfer_id: FB1E8AC6-4FA4-C883-ED5A-54960E88F5FE
          workflow_id: 77EB7E77-465C-FCC6-CEC6-11F6C8938D24
        - aip_id: 2C91774A-6D4A-67B6-DC43-6AED849DACC1
          completed_at: "1997-03-16T23:52:00Z"
          created_at: "1992-11-05T20:18:50Z"
          id: 11298461603816854458
          name: Accusantium dolor molestias.
          original_id: Nihil recusandae.
          pipeline_id: 773F4E7F-A748-DF6D-4355-233071D2534A
          run_id: 74C53540-E974-ABFF-2565-6BF99F9017B2
          started_at: "2008-06-17T22:56:20Z"
          status: abandoned
          transfer_id: FB1E8AC6-4FA4-C883-ED5A-54960E88F5FE
          workflow_id: 77EB7E77-465C-FCC6-CEC6-11F6C8938D24
        next_cursor: Explicabo est aliquid.
      required:
      - items
...